Data.Type.Equality.Hetero

Description

This module shims kind heterogeneous propositional equality.

Note: some instances: Read, Enum, Bounded and Data would be available only since GHC-8.0 as we need ~~ constraint. Also GH-7.10.3 is nitpicky data constructor ‘HRefl’ cannot be GADT-like in its *kind* arguments, thus this module is available only with GHC >= 8.0

Documentation

data (a :: k1) :~~: (b :: k2) where infix 4

Kind heterogeneous propositional equality. Like :~:, a :~~: b is inhabited by a terminating value if and only if a is the same type as b.

Since: base-4.10.0.0

Constructors

HRefl :: forall {k1} (a :: k1). a :~~: a